+++ demos/auth-playground/src/components/AppPinDialog.tsx
@@ -13,6 +13,7 @@ import {
   IonRow,
   IonTitle,
   IonToolbar,
+  useIonModal,
   useIonViewWillEnter,
 } from '@ionic/react';
 import { close } from 'ionicons/icons';
@@ -204,4 +205,24 @@ const AppPinDialog: React.FC<Props> = ({ setPasscodeMode, dismiss }) => {
     </IonPage>
   );
 };
-export default AppPinDialog;
+
+export const usePinDialog = () => {
+  const [isPasscodeSetRequest, setIsPasscodeSetRequest] = useState(false);
+  const [presentPinDialog, dismissPinDialog] = useIonModal(AppPinDialog, {
+    setPasscodeMode: isPasscodeSetRequest,
+    dismiss: (data?: any, role?: string) => dismissPinDialog(data, role),
+  });
+
+  const onPasscodeRequested = async (isPasscodeSetRequest: boolean): Promise<string> => {
+    setIsPasscodeSetRequest(isPasscodeSetRequest);
+    const data = await new Promise<any>((resolve) => {
+      presentPinDialog({
+        backdropDismiss: false,
+        onDidDismiss: (event) => resolve(event.detail.data),
+      });
+    });
+    return data || '';
+  };
+
+  return onPasscodeRequested;
+};
